+++ app/policies/comment_vote_policy.rb
@@ -1,7 +1,10 @@
 class CommentVotePolicy < ApplicationPolicy
+  def create?
+    unbanned? && !record.comment.is_deleted?
+  end
+
   def destroy?
-    # XXX permissions are checked in Comment#unvote!
-    true
+    record.user_id == user.id
   end
 
   def can_see_votes?