+++ test/unit/comment_test.rb
@@ -143,55 +143,6 @@ class CommentTest < ActiveSupport::TestCase
         end
       end
 
-      should "not record the user id of the voter" do
-        user = FactoryBot.create(:user)
-        post = FactoryBot.create(:post)
-        c1 = FactoryBot.create(:comment, :post => post)
-        CurrentUser.scoped(user, "127.0.0.1") do
-          c1.vote!("up")
-          c1.reload
-          assert_not_equal(user.id, c1.updater_id)
-        end
-      end
-
-      should "not allow duplicate votes" do
-        user = FactoryBot.create(:user)
-        post = FactoryBot.create(:post)
-        c1 = FactoryBot.create(:comment, :post => post)
-
-        assert_nothing_raised { c1.vote!("down") }
-        exception = assert_raises(ActiveRecord::RecordInvalid) { c1.vote!("down") }
-        assert_equal("Validation failed: You have already voted for this comment", exception.message)
-        assert_equal(1, CommentVote.count)
-        assert_equal(-1, CommentVote.last.score)
-
-        c2 = FactoryBot.create(:comment, :post => post)
-        assert_nothing_raised { c2.vote!("down") }
-        assert_equal(2, CommentVote.count)
-      end
-
-      should "not allow upvotes by the creator" do
-        user = FactoryBot.create(:user)
-        post = FactoryBot.create(:post)
-        c1 = create(:comment, post: post, creator: CurrentUser.user)
-
-        exception = assert_raises(ActiveRecord::RecordInvalid) { c1.vote!("up") }
-        assert_equal("Validation failed: You cannot upvote your own comments", exception.message)
-      end
-
-      should "allow undoing of votes" do
-        user = FactoryBot.create(:user)
-        post = FactoryBot.create(:post)
-        comment = FactoryBot.create(:comment, :post => post)
-        CurrentUser.scoped(user, "127.0.0.1") do
-          comment.vote!("up")
-          comment.unvote!
-          comment.reload
-          assert_equal(0, comment.score)
-          assert_nothing_raised {comment.vote!("down")}
-        end
-      end
-
       should "be searchable" do
         c1 = FactoryBot.create(:comment, :body => "aaa bbb ccc")
         c2 = FactoryBot.create(:comment, :body => "aaa ddd")